#31 feat/error 상태 추가및 placeholder 스타일 지정
